Add Flake8 at requirements
As any project, shorty should have a linter in order to maintain the
code quality. So we added the flake8 as a requirement.

We can run the code style check by doing:
`flake8 <folder-to-check>`

If we dont give a folder to check the flake8 will scan the working
directory.

Finally it is suggested to exclude the venv folder by adding after
the flake8 comman the following option:
`--exclude=<virtual-env-folder>`

More about flake8:
http://flake8.pycqa.org