Create Functional Tests for Shorty's Api
Remove test_example.py as it is no longer used

Create functional tests that checks all the possible scenarios in our
api. As for the tests' assert we should check the desired responses that
will be implemented soon. In the conftest file, create a post method in
order to execute post request with the test client.

To run the test suite, go to the project folder from terminal
and execute:

$ py.test

More for py.test: https://docs.pytest.org/

Our api have two kind of response object that are based on the JSON:API
specification (v1.0).

The format of the success response object is:

{
  "data": {
    "url": "https://example.com",
    "link": "https://bit.ly/8h1bka"
  }
}

The format of the error response object is:

{
  "error": {
    "status": "the HTTP status code applicable to this problem",
    "code": "an application-specific error code",
    "detail": "a human-readable explanation"
  }
}

More info for JSON:API specification here:
https://jsonapi.org/format/