Create Provider for shorten use case
Create providers that execute the shorten request. The current providers
is Bit.ly and TinyUrl. Each of the providers send a request to the
real provider's api.

If the request is successful, they return the shorten link. Otherwise,
they throw a custom exception that is called ProviderException and
indicates that there was an issue with the selected provider.