Create ShortenUseCase
Create the shorten use case that takes a shorten request, checks if it
is valid and if it is, sends the request to the appropriate provider.
If on the request the provider was not given, the shorten use case
chooses the default shortening provider.

In the future this can be enhanced by adding a logic that if the
selected provider failed, then we should try with the rest of the
providers until on succeeds.

Also enhance the api class to create a new shorten use case for each
request, and if an error occurs return the message with the valid
format.